fix db performance issues

+++ packages/sdc2-server/db/migrations/00000000000000_initialize_schema.js
@@ -21,7 +21,7 @@ exports.up = async function (knex) {
     table.decimal('maximum', 13, 4)
 
     table.unique(['measurementDay', 'location', 'type'], {
-      indexName: 'dailyAggregation_measurementDay_location_type_uindex',
+      indexName: 'daily_aggregation_measurement_day_location_type_uindex',
     })
   })
 
@@ -37,7 +37,7 @@ exports.up = async function (knex) {
     table.decimal('maximum', 13, 4)
 
     table.unique(['measurementHour', 'location', 'type'], {
-      indexName: 'hourlyAggregation_measurementHour_location_type_uindex',
+      indexName: 'hourly_aggregation_measurement_hour_location_type_uindex',
     })
   })
 
@@ -48,17 +48,16 @@ exports.up = async function (knex) {
     table.string('type', 64).notNullable()
     table.decimal('value', 13, 4).notNullable()
 
-    table.index(
-      ['createdAt', 'location', 'type'],
-      'measurement_location_createdAt_type_index'
-    )
     table.index(['location', 'type'], 'measurement_location_type_index')
+    table.index(['createdAt'], 'measurement_created_at_index')
   })
 }
 
 exports.down = async function (knex) {
-  await knex.schema.dropTable('apiKey')
-  await knex.schema.dropTable('dailyAggregation')
-  await knex.schema.dropTable('hourlyAggregation')
-  await knex.schema.dropTable('measurement')
+  if (process.env.NODE_ENV !== 'production') {
+    await knex.schema.dropTable('apiKey')
+    await knex.schema.dropTable('dailyAggregation')
+    await knex.schema.dropTable('hourlyAggregation')
+    await knex.schema.dropTable('measurement')
+  }
 }
